Fixed bugs; Added new caching system
